+++ src/Form/LegacyConsumer/Commands/SetupNewTemplateHook.php
@@ -21,12 +21,12 @@ class SetupNewTemplateHook implements HookCommandInterface {
 		// On the old hook, run the new hook and render the fields.
 		add_action(
 			"give_$hook",
-			static function ( $formID ) use ( $hook ) {
+			static function ( $formId ) use ( $hook ) {
 				$collection = Group::make( $hook );
-				do_action( "give_fields_$hook", $collection, $formID );
+				do_action( "give_fields_$hook", $collection, $formId );
 				$collection->walk(
-					static function ( $node ) use ( $formID ) {
-						FieldView::render( $node, $formID );
+					static function ( $node ) use ( $formId ) {
+						FieldView::render( $node, $formId );
 					}
 				);
 			}